requirements.txt
flask
transformers==4.40.1
torch==2.3.0
Pillow
semantic-kernel==1.29.0
